Move plan properties over to its own file
